【功能修改】: 日历功能完善-8

+++ api/blog-file/src/main/java/com/blog/file/service/CalendarService.java
@@ -35,7 +35,7 @@ public interface CalendarService {
     /**
      * 删除
      */
-    boolean removeById(Long id);
+    boolean removeById(Integer id);
 
     /**
      * 查询指定月份的全部日历记录
